Builds in windows and links. Unresolved: how to convince the ogregraphics.dll that the static sInstance of the plugin registration singleton lives in libcore instead of locally

+++ libcore/src/transfer/TransferData.hpp
@@ -273,9 +273,9 @@ public:
 			if (data == NULL && length == 0) {
 				break;
 			} else if (data == NULL) {
-				context.updateZeros(length);
+				context.updateZeros((size_t)length);
 			} else {
-				context.update(data, length);
+				context.update(data, (size_t)length);
 			}
 		}
 		return context.get();
